+++ tools/otp_enroll.py
@@ -75,7 +75,7 @@ def main():
 	parser.epilog = PARSER_EPILOG
 	arguments = parser.parse_args()
 
-	utilities.configure_stream_logger(arguments.loglvl, arguments.logger)
+	utilities.configure_stream_logger(arguments.logger, arguments.loglvl)
 
 	if arguments.database_url:
 		database_connection_url = arguments.database_url