+++ docs/source/change_log.rst
@@ -14,6 +14,7 @@ Version 1.6.0
 
 * Support negotiating STARTTLS with SMTP servers that support it
 * Support for real time event publishing to the client
+* More flexibility in configuring server logging
 
 Version 1.5.1
 ^^^^^^^^^^^^^